Do not write data to db if data[:name] == 'Vaccant Poste'
Posts vacated by members are listed on the site. Their URLs haven't been
replace and as we form the member IDs from the URLs, we are reassigning
upstream IDs of members to new members name 'Vacant Poste'. We don't
want to do that, so we skip writing the data to the database.